fix(webhooks): Refactor webhook event tasks
