+++ cl/api/tasks.py
@@ -1,0 +1,72 @@
+import json
+
+from rest_framework.renderers import JSONRenderer
+
+from cl.api.models import Webhook, WebhookEvent, WebhookEventType
+from cl.api.utils import generate_webhook_key_content
+from cl.api.webhooks import send_webhook_event
+from cl.celery_init import app
+from cl.corpus_importer.api_serializers import DocketEntrySerializer
+from cl.search.models import DocketEntry
+
+
+@app.task()
+def send_test_webhook_event(
+    webhook_pk: int,
+    content_str: str,
+) -> None:
+    """POSTS the test webhook event.
+
+    :param webhook_pk: The webhook primary key.
+    :param content_str: The str content to POST.
+    :return: None
+    """
+
+    webhook = Webhook.objects.get(pk=webhook_pk)
+    json_obj = json.loads(content_str)
+    webhook_event = WebhookEvent.objects.create(
+        webhook=webhook, content=json_obj, debug=True
+    )
+    send_webhook_event(webhook_event, content_str.encode("utf-8"))
+
+
+@app.task()
+def send_docket_alert_webhook_events(
+    des_pks: list[int],
+    webhook_recipients_pks: list[int],
+) -> None:
+    """POSTS the DocketAlert to the recipients webhook(s)
+
+    :param des_pks: The list of docket entries primary keys.
+    :param webhook_recipients_pks: A list of User pks to send the webhook to.
+    :return: None
+    """
+
+    webhooks = Webhook.objects.filter(
+        event_type=WebhookEventType.DOCKET_ALERT,
+        user_id__in=webhook_recipients_pks,
+        enabled=True,
+    )
+    docket_entries = DocketEntry.objects.filter(pk__in=des_pks)
+    serialized_docket_entries = []
+    for de in docket_entries:
+        serialized_docket_entries.append(DocketEntrySerializer(de).data)
+
+    for webhook in webhooks:
+        post_content = {
+            "webhook": generate_webhook_key_content(webhook),
+            "payload": {
+                "results": serialized_docket_entries,
+            },
+        }
+        renderer = JSONRenderer()
+        json_bytes = renderer.render(
+            post_content,
+            accepted_media_type="application/json;",
+        )
+
+        webhook_event = WebhookEvent.objects.create(
+            webhook=webhook,
+            content=post_content,
+        )
+        send_webhook_event(webhook_event, json_bytes)
